UPDATE: made a hopefully more robust set of release targets
make release-dev
make release-test
make release-minor
make release-major

These check that everything has been committed, and that we are on master.   If so compile the requirements.txt, commit it to git, bumpversion with the appropriate target from the suffix, and do bin/release.sh

We still have to build the container again and push it up to dockerhub afterwards.